feat(workspace): inject bound-skill prompts into task runs
Task/orchestration runs only delivered skill content on the chat path, so
CLI agents executing tasks got the task description but none of their bound
skills' instructions — they would hunt the filesystem and fail. Mirror the
chat path in the task path.

- move appendSkillPromptsFromResolved + its config helpers out of chathttp
  into internal/workspace (where ResolvedSkill lives) as exported
  AppendSkillPromptsFromResolved, so chat and task share one implementation
- carry EffectiveSkills through resolvedTaskAgent and append them to the
  task system prompt; "only-when-bound" is automatic (skill-less agents add
  nothing, so there is no extra cost)
- chathttp now calls workspace.AppendSkillPromptsFromResolved

Adds skill_prompt_test.go covering append, the no-skills no-op, and the
planning-profile binding-settings rendering.
