+++ internal/workspace/task_handler_agent_resolver.go
@@ -29,8 +29,9 @@ func (h *LLMTaskHandler) resolveExecutionAgent(agentName string, task Task) (*re
 		}
 		if resolved != nil && resolved.Agent != nil {
 			return &resolvedTaskAgent{
-				Agent:      resolved.Agent,
-				MCPServers: append([]string{}, resolved.MCPServers...),
+				Agent:           resolved.Agent,
+				MCPServers:      append([]string{}, resolved.MCPServers...),
+				EffectiveSkills: append([]ResolvedSkill{}, resolved.EffectiveSkills...),
 			}, nil
 		}
 	}
